sort and norm
