+++ src/regex/items/brackets.c
