+++ src/regex/items/factory.c
