+++ src/regex/items/manager.c
