+++ src/regex/items/utils.c
