+++ src/research.c
@@ -1,4 +1,14 @@
-
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   research.c                                         :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: lroussel <lroussel@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/04/03 16:48:04 by lroussel          #+#    #+#             */
+/*   Updated: 2025/04/03 16:48:06 by lroussel         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
 
 #include "regex.h"
